refactor: extract components into own files
